# Convert Imports to Indirect Calls for Dynamic Linking