This complements the work done in https://github.com/WebAssembly/binaryen/pull/2523 to convert Imports in Dynamic Linking into Indirect Calls. The benefits of doing this are:

1. Faster performance
2. Skip legalization
3. Remove JS stubs (To be done in a later PR)

A sample benchmark has been uploaded here: https://github.com/awtcode/dylink_benchmark

These are the results of the benchmark:

| Chrome 78  |  |
| ------------- | ------------- |
| Original  | 227ms  |
| PR  | 17ms  |

| Firefox 71  |  |
| ------------- | ------------- |
| Original  | 43ms  |
| PR  | 6ms  |

This PR gives a 13x improvement in Chrome and a 7x improvement in FF.