@awtcode 2019-12-11
@sbc100 , this is my initial attempt to replace imports with indirect calls. Can you take a look and give me your comments? Thanks.