@awtcode 2019-12-12
@sbc100 and @kripken , I have updated the PR with some results from my test app.