@kripken 2019-12-12
I am uncomfortable with getting the list of JS library functions in this way. It means that there are two kinds of imports, one that is provided by wasm and one by JS, and the module being compiled needs to know which is which at compile time. This feels a little risky.

IIUC this PR makes imported wasm functions be called indirectly. If that is fast enough, then making the JS functions be indirect calls as well may be fast enough too. That is, JS library functions would be added to the table, then treated just like any other import. That layer could also add legalization for JS as needed, which would simplify our legalization story.

(However, that assumes these indirect calls are fast enough, and we may want to keep one direction's calls as direct, for speed? That would make this more complicated.)